Add a new dispatcher for transcode workers
This has a smaller scope than transcodeChanges.sh and is therefore simpler and
more robust.
It solves the particular issue of being able to scale the number of workers on
each machine by virtue of not caring about more than one machine.
Instead to spread the load an instance is run for each machine.

A dummy worker is included for test purposes.